% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapValuesScript.R
\name{mapValues}
\alias{mapValues}
\title{Make a USA map with all the counties colored in}
\usage{
mapValues(df, county_index = 1, plot_index = 2, save = TRUE,
  filename = "./usa_plot.png", titleString = "default plot")
}
\arguments{
\item{df}{The data frame of the counties in the US and the data to plot}

\item{county_index}{The numerical index (base 1) of the column which contains
the FIPS code}

\item{plot_index}{The numerical index (base 1) of the column which contains
the data we want to plot}

\item{save}{T/F determines if we want to save the plot}

\item{filename}{If we want to save, where are we saving?}
}
\description{
Make a USA map with all the counties colored in
}
